LayoutTests/js/dfg-make-rope-side-effects-expected.txt

Check that the DFG can handle MakeRope on values that have side effects on [[ToPrimitive]].

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testOutput.length is 0
PASS successfullyParsed is true

TEST COMPLETE
